fixed scale bug

+++ Engine/ressources/template/engine/src/manager/initializer.cpp
@@ -196,7 +196,7 @@ void s2d::Initializer::initAnimation(const std::string& path, s2d::SpriteReposit
 			std::vector<std::string> propertys = std::splitString(line, DELIMITER);
 
 			if(propertys[1] != "")
-				frames.push_back(s2d::KeyFrame(propertys[1], atoi(propertys[0].c_str())));
+				frames.push_back(s2d::KeyFrame(propertys[1], std::stof(propertys[0].c_str())));
 		}
 
 	}
