+++ Engine/ressources/template/engine/src/spriteComponents/keyFrame.h
@@ -7,11 +7,11 @@ namespace s2d
 	struct KeyFrame
 	{
 		std::string path;
-		int delay;
-		int position;
+		float delay;
+		uint32_t position;
 
 		KeyFrame() { this->path = ""; this->delay = 0; }
-		KeyFrame(const std::string& path, const int delay)
+		KeyFrame(const std::string& path, const float delay)
 		{
 			this->position = -1;
 			this->path = path;
